Move AddRateModal and EditRateModal files.
